Modified events
First something  creates & registers the event (if there is only one
thing that thing creates & registers it)
Then when the activator activates the event. The events class finds the
event and activates it with the Object input. Then the event that was
found transforms it into the specific thing that it sends (i.e. Key
event)